+++ training_log_app/src/entities/log_entry.py
@@ -7,15 +7,14 @@ class LogEntry:
     After creation more info can be added.
 
     Attributes:
+        user_id: type int
         date: type string
         duration: type integer in minutes
-        session_style: type string   e.g. boxing, wrestling, perhaps multiple choice?
+        session_style: type string   e.g. boxing, wrestling
         what_went_well: type string
         what_did_not_go_well: type string
         goal_for_next_session: type string
         was_last_goal_achieved: type int, 1 for True, 0 for False
-        user_id: type int
-        log_entry_id: type int created in database
     '''
 
     # I believe this is the simplest way to construct an entry so pylint is disabled
@@ -30,6 +29,3 @@ class LogEntry:
         self.goal_for_next_session = None
         self.was_last_goal_achieved = None
         self.log_id = None
-
-    def add_id(self, id):
-        self.log_id = id