+++ training_log_app/src/ui/ui.py
@@ -4,6 +4,7 @@ from ui.login_view import LoginView
 from ui.create_new_user_view import CreateNewUserView
 from ui.main_user_page_view import MainUserPageView
 from ui.statistics_view import StatisticsPageView
+from ui.browse_log_entries_view import BrowseLogEntriesView
 
 
 class UI:
@@ -22,7 +23,7 @@ class UI:
 
     def configure_styles(self):
         # configure global ui styling
-        ttk.Style().configure('.', font=('Times New Roman', 12), foreground='black')
+        ttk.Style().configure('.', font=('Khmer OS', 11))
 
     def _hide_current_view(self):
         if self._current_view:
@@ -55,7 +56,8 @@ class UI:
         self._current_view = MainUserPageView(self._root,
                                               self._show_login_view,
                                               self._show_new_log_entry_view,
-                                              self._show_statistics_page_view)
+                                              self._show_statistics_page_view,
+                                              self._show_browse_log_entries_view)
         self._current_view.pack()
 
     def _show_create_new_user_view(self):
@@ -71,5 +73,13 @@ class UI:
 
         self._current_view = StatisticsPageView(self._root,
                                                 self._show_main_user_page)
+
+        self._current_view.pack()
+
+    def _show_browse_log_entries_view(self):
+        self._hide_current_view()
+
+        self._current_view = BrowseLogEntriesView(self._root,
+                                                  self._show_main_user_page)
 
         self._current_view.pack()